+++ Cargo.toml
@@ -37,7 +37,6 @@ members = [
     "src/query/functions",
     "src/query/functions-v2",
     "src/query/management",
-    "src/query/planner",
     "src/query/pipeline/core",
     "src/query/pipeline/sinks",
     "src/query/pipeline/sources",
