--- src/query/planner/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "common-planner"
-version = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-publish = { workspace = true }
-edition = { workspace = true }
-
-[dependencies]
-common-datavalues = { path = "../datavalues" }
-common-exception = { path = "../../common/exception" }
-common-meta-app = { path = "../../meta/app" }
-common-meta-types = { path = "../../meta/types" }
-
-once_cell = "1.15.0"
-parking_lot = "0.12"
-typetag = "0.2.3"
-
-serde = { workspace = true }
-serde_json = { workspace = true }
