--- src/query/planner/src/plans/mod.rs
@@ -1,37 +1,0 @@
-// Copyright 2022 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod alter_udf;
-mod alter_view;
-mod call;
-mod create_udf;
-mod create_view;
-mod delete;
-mod drop_udf;
-mod drop_view;
-mod list;
-mod projection;
-mod setting;
-
-pub use alter_udf::AlterUDFPlan;
-pub use alter_view::AlterViewPlan;
-pub use call::CallPlan;
-pub use create_udf::CreateUDFPlan;
-pub use create_view::CreateViewPlan;
-pub use delete::*;
-pub use drop_udf::DropUDFPlan;
-pub use drop_view::DropViewPlan;
-pub use list::ListPlan;
-pub use projection::*;
-pub use setting::*;
