+++ src/query/sql/src/planner/plans/mod.rs
@@ -13,53 +13,69 @@
 // limitations under the License.
 
 mod aggregate;
+mod call;
 mod copy_v2;
 mod ddl;
+mod delete;
 mod dummy_table_scan;
 mod eval_scalar;
 mod exchange;
+mod extras;
 mod filter;
 mod hash_join;
 pub mod insert;
 mod kill;
 mod limit;
+mod list;
 mod logical_get;
 mod logical_join;
 mod operator;
+mod partition;
 mod pattern;
 mod physical_scan;
 mod plan;
+mod plan_read_datasource;
 mod presign;
+mod projection;
 mod recluster_table;
 mod scalar;
+mod setting;
 pub mod share;
 mod sort;
 mod union_all;
 mod update;
 
 pub use aggregate::*;
+pub use call::CallPlan;
 pub use copy_v2::*;
 pub use ddl::*;
+pub use delete::DeletePlan;
 pub use dummy_table_scan::DummyTableScan;
 pub use eval_scalar::*;
 pub use exchange::*;
+pub use extras::*;
 pub use filter::*;
 pub use hash_join::PhysicalHashJoin;
 pub use insert::Insert;
 pub use insert::InsertInputSource;
 pub use kill::KillPlan;
 pub use limit::*;
+pub use list::ListPlan;
 pub use logical_get::*;
 pub use logical_join::*;
 pub use operator::*;
+pub use partition::*;
 pub use pattern::PatternPlan;
 pub use physical_scan::PhysicalScan;
 pub use plan::Plan::*;
 pub use plan::RewriteKind::*;
 pub use plan::*;
+pub use plan_read_datasource::*;
 pub use presign::*;
+pub use projection::Projection;
 pub use recluster_table::ReclusterTablePlan;
 pub use scalar::*;
+pub use setting::*;
 pub use share::*;
 pub use sort::*;
 pub use union_all::UnionAll;
