+++ src/query/storages/hive/hive/Cargo.toml
@@ -23,7 +23,6 @@ common-meta-app = { path = "../../../../meta/app" }
 common-meta-types = { path = "../../../../meta/types" }
 common-pipeline-core = { path = "../../../pipeline/core" }
 common-pipeline-sources = { path = "../../../pipeline/sources" }
-common-planner = { path = "../../../planner" }
 common-storage = { path = "../../../../common/storage" }
 common-storages-cache = { path = "../../cache" }
 common-storages-index = { path = "../../index" }
