+++ src/query/storages/index/Cargo.toml
@@ -20,7 +20,6 @@ common-datablocks = { path = "../../datablocks" }
 common-datavalues = { path = "../../datavalues" }
 common-exception = { path = "../../../common/exception" }
 common-functions = { path = "../../functions" }
-common-planner = { path = "../../planner" }
 common-sql = { path = "../../sql" }
 common-storages-table-meta = { path = "../table-meta" }
