+++ src/query/storages/system/Cargo.toml
@@ -26,7 +26,6 @@ common-pipeline-core = { path = "../../pipeline/core" }
 common-pipeline-sinks = { path = "../../pipeline/sinks" }
 common-pipeline-sources = { path = "../../pipeline/sources" }
 common-pipeline-transforms = { path = "../../pipeline/transforms" }
-common-planner = { path = "../../planner" }
 common-storage = { path = "../../../common/storage" }
 common-storages-view = { path = "../view" }
 common-users = { path = "../../users" }
